change a type cast in Function::cnfs() from int to long for compliance with 64bit compiler; rename homework3.cpp to homework.cpp

+++ AST_class.h
@@ -245,7 +245,7 @@ CNFs *Function::cnfs()
 {
 	ARG_LIST args;
 	for (std::list<Expression*>::iterator iter = arg_list.begin(); iter != arg_list.end(); ++iter)
-		args.args.push_back((int)((*iter)->cnfs()));
+		args.args.push_back((long)((*iter)->cnfs()));
 	int _id = id;
 	if (op == NOT)
 		_id = ~_id;
